fix: fix generation of changelog (#17)

+++ scripts/update-readme-version.sh
@@ -39,22 +39,25 @@ print_info "Updating version references in $README_FILE to $VERSION"
 cp "$README_FILE" "${README_FILE}.bak"
 
 # Update go get command (add version if not present, update if present)
-sed -i.bak "s/go get github.com\/ignatij\/goflow\(@v[0-9]\+\.[0-9]\+\.[0-9]\+\)\?/go get github.com\/ignatij\/goflow@v$VERSION/g" "$README_FILE"
+sed -i.bak "s/go get github.com\/ignatij\/goflow$/go get github.com\/ignatij\/goflow@v$VERSION/g" "$README_FILE"
+sed -i.bak "s/go get github.com\/ignatij\/goflow@v[0-9]\+\.[0-9]\+\.[0-9]\+/go get github.com\/ignatij\/goflow@v$VERSION/g" "$README_FILE"
 
 # Update require statement in go.mod example
-sed -i.bak "s/require github.com\/ignatij\/goflow v[0-9]\+\.[0-9]\+\.[0-9]\+/require github.com\/ignatij\/goflow v$VERSION/g" "$README_FILE"
+sed -i.bak "s/v[0-9]\+\.[0-9]\+\.[0-9]\+/v$VERSION/g" "$README_FILE"
 
 # Update any version badges (if they exist)
 sed -i.bak "s/badge\/version-[0-9]\+\.[0-9]\+\.[0-9]\+/badge\/version-$VERSION/g" "$README_FILE"
 
-# Remove backup files
+# Remove backup files created by sed
 rm -f "${README_FILE}.bak"
 
 print_success "Updated $README_FILE with version $VERSION"
 
 # Show what was changed
 print_info "Changes made:"
-git diff --no-index "${README_FILE}.bak" "$README_FILE" || true
-
-# Clean up backup
-rm -f "${README_FILE}.bak" 
+if [ -f "${README_FILE}.bak" ]; then
+    git diff --no-index "${README_FILE}.bak" "$README_FILE" || true
+    rm -f "${README_FILE}.bak"
+else
+    print_info "No changes detected (version references may already be up to date)"
+fi
